# gestimowebfront-end/src/tools/swagger/src/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Swagger API pour GESTIMOWEB REST API",
    "version" : "0.0.1-SNAPSHOT",
    "title" : "Swagger GESTIMOWEB REST API",
    "contact" : {
      "name" : "bzadata",
      "email" : "michel.bossoh@outlook.fr"
    }
  },
  "basePath" : "http://localhost:8282/",
  "paths" : {
...
    "gestimoweb/api/v1/appartement/all/{id}" : {
      "get" : {
        "operationId" : "findAllAppartement",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/AppartementDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/appartement/alllmeublelibre/{idAgence}" : {
      "get" : {
        "operationId" : "findAllAppartementLibre",
        "parameters" : [ {
          "name" : "idAgence",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/AppartementDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/appartement/allmeuble/{id}" : {
      "get" : {
        "operationId" : "findAllAppartementMeuble",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/AppartementDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/appartement/delete/{id}" : {
      "delete" : {
        "operationId" : "deleteAppartement",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/appartement/findById/{id}" : {
      "get" : {
        "operationId" : "findByIDAppartement",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/AppartementDto"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/appartement/findByIdEtage/{id}" : {
      "get" : {
        "operationId" : "findByIdEtageAppartement",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/AppartementDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/appartement/findByName/{name}" : {
      "get" : {
        "operationId" : "findByNameAppartement",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/AppartementDto"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/appartement/findallappartementbyidcategorie/{id}" : {
      "get" : {
        "operationId" : "findAllAppartementByIdCategorie",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/AppartementDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/appartement/save" : {
      "post" : {
        "operationId" : "saveAppartement",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/AppartementDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/AppartementDto"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/appartement/saveForCategorie" : {
      "post" : {
        "operationId" : "saveForCategorie",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/AppartementDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/AppartementDto"
            }
          }
        }
      }
    },
...
    "gestimoweb/api/v1/bail/save" : {
      "post" : {
        "operationId" : "modifierUnBail",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/BailModifDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/OperationDto"
            }
          }
        }
      }
    },
...
    "gestimoweb/api/v1/bailmagasin/all/{idAgence}" : {
      "get" : {
        "operationId" : "findAllBailMagasin",
        "parameters" : [ {
          "name" : "idAgence",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/BailMagasinDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/bailmagasin/delete/{id}" : {
      "delete" : {
        "operationId" : "deleteBailMagasin",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/bailmagasin/findById/{id}" : {
      "get" : {
        "operationId" : "findByIDBailMagasin",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/BailMagasinDto"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/bailmagasin/findByName/{name}" : {
      "get" : {
        "operationId" : "findByNameBailMagasin",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/BailMagasinDto"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/bailmagasin/save" : {
      "post" : {
        "operationId" : "saveBailMagasin",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/BailMagasinDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/OperationDto"
            }
          }
        }
      }
    },
...
    "gestimoweb/api/v1/magasin/all/{idAgence}" : {
      "get" : {
        "operationId" : "findAllMagasin",
        "parameters" : [ {
          "name" : "idAgence",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/MagasinResponseDto"
              }
            }
          }
        }
      }
    },
...
    "gestimoweb/api/v1/magasin/findByName/{name}" : {
      "get" : {
        "operationId" : "findByNameMagasinDto",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/MagasinDto"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "AppartementDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "nbrPieceApp" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nbreChambreApp" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nbreSalonApp" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nbreSalleEauApp" : {
          "type" : "integer",
          "format" : "int32"
        },
        "numApp" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idEtageAppartement" : {
          "type" : "integer",
          "format" : "int64"
        },
        "fullNameProprio" : {
          "type" : "string"
        },
        "codeAbrvBienImmobilier" : {
          "type" : "string"
        },
        "nomCompletBienImmobilier" : {
          "type" : "string"
        },
        "nomBaptiserBienImmobilier" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "superficieBien" : {
          "type" : "number",
          "format" : "double"
        },
        "bienMeublerResidence" : {
          "type" : "boolean"
        },
        "nameCategorie" : {
          "type" : "string"
        },
        "priceCategorie" : {
          "type" : "number",
          "format" : "double"
        },
        "nbrDiffJourCategorie" : {
          "type" : "integer",
          "format" : "int32"
        },
        "pourcentReducCategorie" : {
          "type" : "number",
          "format" : "double"
        },
        "idCategorieChambre" : {
          "$ref" : "#/definitions/CategoryChambreSaveOrUpdateDto"
        },
        "idChapitre" : {
          "type" : "integer",
          "format" : "int64"
        },
        "occupied" : {
          "type" : "boolean"
        }
      }
    },
...
    "BailMagasinDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "designationBail" : {
          "type" : "string"
        },
        "abrvCodeBail" : {
          "type" : "string"
        },
        "enCoursBail" : {
          "type" : "boolean"
        },
        "archiveBail" : {
          "type" : "boolean"
        },
        "montantCautionBail" : {
          "type" : "number",
          "format" : "double"
        },
        "nbreMoisCautionBail" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nouveauMontantLoyer" : {
          "type" : "number",
          "format" : "double"
        },
        "dateDebut" : {
          "type" : "string",
          "format" : "date"
        },
        "dateFin" : {
          "type" : "string",
          "format" : "date"
        },
        "idMagasin" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idLocataire" : {
          "type" : "integer",
          "format" : "int64"
        },
        "nomLocataire" : {
          "type" : "string"
        },
        "codeBien" : {
          "type" : "string"
        }
      }
    },
    "BailModifDto" : {
      "type" : "object",
      "properties" : {
        "idBail" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "nombreMoisCaution" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nouveauMontantLoyer" : {
          "type" : "number",
          "format" : "double"
        },
        "ancienMontantLoyer" : {
          "type" : "number",
          "format" : "double"
        },
        "dateDePriseEncompte" : {
          "type" : "string",
          "format" : "date"
        },
        "dateFin" : {
          "type" : "string",
          "format" : "date"
        }
      }
    },
...
    "CategoryChambreSaveOrUpdateDto" : {
      "type" : "object",
      "required" : [ "description", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "prixGategorieDto" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/PrixParCategorieChambreDto"
          }
        },
        "appartements" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/AppartementDto"
          }
        }
      }
    },
...
    "MagasinDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "nombrePieceMagasin" : {
          "type" : "integer",
          "format" : "int32"
        },
        "numMagasin" : {
          "type" : "integer",
          "format" : "int64"
        },
        "codeAbrvBienImmobilier" : {
          "type" : "string"
        },
        "nomCompletBienImmobilier" : {
          "type" : "string"
        },
        "nomBaptiserBienImmobilier" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "superficieBien" : {
          "type" : "number",
          "format" : "double"
        },
        "bienMeublerResidence" : {
          "type" : "boolean"
        },
        "idEtage" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idSite" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idUtilisateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "proprietaire" : {
          "type" : "string"
        },
        "idmmeuble" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idChapitre" : {
          "type" : "integer",
          "format" : "int64"
        },
        "underBuildingMagasin" : {
          "type" : "boolean"
        },
        "occupied" : {
          "type" : "boolean"
        }
      }
    },
    "MagasinResponseDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "nombrePieceMagasin" : {
          "type" : "integer",
          "format" : "int32"
        },
        "numMagasin" : {
          "type" : "integer",
          "format" : "int64"
        },
        "codeAbrvBienImmobilier" : {
          "type" : "string"
        },
        "nomCompletBienImmobilier" : {
          "type" : "string"
        },
        "nomBaptiserBienImmobilier" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "superficieBien" : {
          "type" : "number",
          "format" : "double"
        },
        "bienMeublerResidence" : {
          "type" : "boolean"
        },
        "proprietaire" : {
          "type" : "string"
        },
        "underBuildingMagasin" : {
          "type" : "boolean"
        },
        "occupied" : {
          "type" : "boolean"
        }
      }
    },
...
    "OperationDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idFirstAppel" : {
          "type" : "integer",
          "format" : "int64"
        },
        "creationDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lastModifiedDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "dateDebut" : {
          "type" : "string",
          "format" : "date"
        },
        "dateFin" : {
          "type" : "string",
          "format" : "date"
        },
        "utilisateurOperation" : {
          "type" : "string"
        },
        "bienImmobilierOperation" : {
          "type" : "string"
        },
        "designationBail" : {
          "type" : "string"
        },
        "abrvCodeBail" : {
          "type" : "string"
        },
        "enCoursBail" : {
          "type" : "boolean"
        },
        "archiveBail" : {
          "type" : "boolean"
        },
        "montantCautionBail" : {
          "type" : "number",
          "format" : "double"
        },
        "nbreMoisCautionBail" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nouveauMontantLoyer" : {
          "type" : "number",
          "format" : "double"
        },
        "idBienImmobilier" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idLocataire" : {
          "type" : "integer",
          "format" : "int64"
        },
        "codeAbrvBienImmobilier" : {
          "type" : "string"
        }
      }
    },
...
    "PrixParCategorieChambreDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "nombreDeJour" : {
          "type" : "string"
        },
        "prix" : {
          "type" : "number",
          "format" : "double"
        },
        "intervalPrix" : {
          "type" : "integer",
          "format" : "int32"
        },
        "description" : {
          "type" : "string"
        },
        "nbrDiffJour" : {
          "type" : "integer",
          "format" : "int32"
        },
        "idCategorieChambre" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
...
  }
}